move: un exported builder methods to another dir
